+++ src/Stack/Options.hs
@@ -1,4 +1,4 @@
-{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE OverloadedStrings,RecordWildCards #-}
 
 module Stack.Options
     (Command(..)
@@ -533,10 +533,8 @@ execOptsExtraParser = eoPlainParser <|>
 globalOptsParser :: Bool -> Parser GlobalOptsMonoid
 globalOptsParser hide0 =
     GlobalOptsMonoid <$>
-    optional (strOption (long Docker.reExecArgName <>
-                         hidden <>
-                         internal <>
-                         hide)) <*>
+    optional (strOption (long Docker.reExecArgName <> hidden <> internal)) <*>
+    optional (option auto (long dockerEntrypointArgName <> hidden <> internal)) <*>
     logLevelOptsParser hide0 <*>
     configOptsParser hide0 <*>
     optional (abstractResolverOptsParser hide0) <*>
@@ -554,14 +552,15 @@ globalOptsParser hide0 =
 
 -- | Create GlobalOpts from GlobalOptsMonoid.
 globalOptsFromMonoid :: Bool -> GlobalOptsMonoid -> GlobalOpts
-globalOptsFromMonoid defaultTerminal gm = GlobalOpts
-    { globalReExecVersion = globalMonoidReExecVersion gm
-    , globalLogLevel = fromMaybe defaultLogLevel (globalMonoidLogLevel gm)
-    , globalConfigMonoid = globalMonoidConfigMonoid gm
-    , globalResolver = globalMonoidResolver gm
-    , globalCompiler = globalMonoidCompiler gm
-    , globalTerminal = fromMaybe defaultTerminal (globalMonoidTerminal gm)
-    , globalStackYaml = globalMonoidStackYaml gm }
+globalOptsFromMonoid defaultTerminal GlobalOptsMonoid{..} = GlobalOpts
+    { globalReExecVersion = globalMonoidReExecVersion
+    , globalDockerEntrypoint = globalMonoidDockerEntrypoint
+    , globalLogLevel = fromMaybe defaultLogLevel (globalMonoidLogLevel)
+    , globalConfigMonoid = globalMonoidConfigMonoid
+    , globalResolver = globalMonoidResolver
+    , globalCompiler = globalMonoidCompiler
+    , globalTerminal = fromMaybe defaultTerminal (globalMonoidTerminal)
+    , globalStackYaml = globalMonoidStackYaml }
 
 initOptsParser :: Parser InitOpts
 initOptsParser =