+++ ChangeLog.md
@@ -5,6 +5,8 @@ Major changes:
 * GHCJS can now be used with stackage snapshots via the new `compiler` field.
 * Windows installers are now available:
   [download them here](https://github.com/commercialhaskell/stack/blob/release/doc/install_and_upgrade.md#windows) [#613](https://github.com/commercialhaskell/stack/issues/613)
+* Docker integration works with non-FPComplete generated images
+  [#531](https://github.com/commercialhaskell/stack/issues/531)
 
 Other enhancements: